# Add vectorized operators