--- a/physicalplan/exchange/coalesce.go
+++ b/physicalplan/exchange/coalesce.go
@@ -11,6 +11,18 @@ import (
 	"github.com/thanos-community/promql-engine/physicalplan/model"
 )
 
+type errorChan chan error
+
+func (c errorChan) getError() error {
+	for err := range c {
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
 type coalesceOperator struct {
 	once   sync.Once
 	series []labels.Labels
@@ -39,6 +51,63 @@ func (c *coalesceOperator) Series(ctx context.Context) ([]labels.Labels, error) 
 	return c.series, nil
 }
 
+func (c *coalesceOperator) Next(ctx context.Context) ([]model.StepVector, error) {
+	var out []model.StepVector = nil
+	var wg sync.WaitGroup
+	var mu sync.RWMutex
+	var errChan = make(errorChan, len(c.operators))
+	for _, o := range c.operators {
+		wg.Add(1)
+		go func(o model.VectorOperator) {
+			defer wg.Done()
+
+			in, err := o.Next(ctx)
+			if err != nil {
+				errChan <- err
+				return
+			}
+			if in == nil {
+				return
+			}
+			mu.RLock()
+			if len(in) > 0 && out == nil {
+				mu.RUnlock()
+				mu.Lock()
+				if len(in) > 0 && out == nil {
+					out = c.pool.GetVectorBatch()
+					for i := 0; i < len(in); i++ {
+						out = append(out, c.pool.GetStepVector(in[i].T))
+					}
+				}
+				mu.Unlock()
+			} else {
+				mu.RUnlock()
+			}
+
+			mu.Lock()
+			for i := 0; i < len(in); i++ {
+				out[i].Samples = append(out[i].Samples, in[i].Samples...)
+				out[i].SampleIDs = append(out[i].SampleIDs, in[i].SampleIDs...)
+				o.GetPool().PutStepVector(in[i])
+			}
+			mu.Unlock()
+			o.GetPool().PutVectors(in)
+		}(o)
+	}
+	wg.Wait()
+	close(errChan)
+
+	if err := errChan.getError(); err != nil {
+		return nil, err
+	}
+
+	if out == nil {
+		return nil, nil
+	}
+
+	return out, nil
+}
+
 func (c *coalesceOperator) loadSeries(ctx context.Context) error {
 	size := 0
 	for i := 0; i < len(c.operators); i++ {
@@ -62,40 +131,7 @@ func (c *coalesceOperator) loadSeries(ctx context.Context) error {
 		}
 	}
 	c.series = result
-	c.pool.SetStepSamplesSize(len(c.series))
+	c.pool.SetStepSize(len(c.series))
 
 	return nil
-}
-
-func (c *coalesceOperator) Next(ctx context.Context) ([]model.StepVector, error) {
-	var out []model.StepVector = nil
-	for _, o := range c.operators {
-		in, err := o.Next(ctx)
-		if err != nil {
-			return nil, err
-		}
-		if in == nil {
-			continue
-		}
-		if len(in) > 0 && out == nil {
-			out = c.pool.GetVectors()
-			for i := 0; i < len(in); i++ {
-				out = append(out, model.StepVector{
-					T:       in[i].T,
-					Samples: c.pool.GetSamples(),
-				})
-			}
-		}
-
-		for i := 0; i < len(in); i++ {
-			out[i].Samples = append(out[i].Samples, in[i].Samples...)
-			o.GetPool().PutSamples(in[i].Samples)
-		}
-		o.GetPool().PutVectors(in)
-	}
-	if out == nil {
-		return nil, nil
-	}
-
-	return out, nil
 }