When aggregating without grouping labels, the aggregation can be done with vectorized assembly operators on the entire step vector.

This commit introduces two aggregation tables as substitute for the old one: scalarTable and vectorizedTable. 
The scalar table uses go code, while the vectorizedTable uses the gonum [floats](https://pkg.go.dev/gonum.org/v1/gonum@v0.12.0/floats) package to execute vectorized aggregations.

Fixes https://github.com/thanos-community/promql-engine/issues/4

<pre>
benchstat old.out new.out 
name                                        old time/op    new time/op    delta
OldEngine/sum/current_engine-4                 123ms ± 0%     122ms ± 0%   ~     (p=1.000 n=1+1)
<strong>OldEngine/sum/new_engine-4                    67.1ms ± 0%    48.9ms ± 0%   ~     (p=1.000 n=1+1)</strong>
OldEngine/sum_rate/current_engine-4            219ms ± 0%     201ms ± 0%   ~     (p=1.000 n=1+1)
<strong>OldEngine/sum_rate/new_engine-4                157ms ± 0%      97ms ± 0%   ~     (p=1.000 n=1+1)</strong>

name                                        old alloc/op   new alloc/op   delta
OldEngine/sum/current_engine-4                4.94MB ± 0%    4.94MB ± 0%   ~     (p=1.000 n=1+1)
<strong>OldEngine/sum/new_engine-4                    18.5MB ± 0%    10.0MB ± 0%   ~     (p=1.000 n=1+1)</strong>
OldEngine/sum_rate/current_engine-4           6.01MB ± 0%    6.01MB ± 0%   ~     (p=1.000 n=1+1)
<strong>OldEngine/sum_rate/new_engine-4               19.0MB ± 0%    12.0MB ± 0%   ~     (p=1.000 n=1+1)</strong>

name                                        old allocs/op  new allocs/op  delta
OldEngine/sum/current_engine-4                 73.8k ± 0%     73.8k ± 0%   ~     (all equal)
OldEngine/sum/new_engine-4                     71.6k ± 0%     73.5k ± 0%   ~     (p=1.000 n=1+1)
OldEngine/sum_rate/current_engine-4            82.9k ± 0%     82.9k ± 0%   ~     (p=1.000 n=1+1)
OldEngine/sum_rate/new_engine-4                98.5k ± 0%    100.4k ± 0%   ~     (p=1.000 n=1+1)
</pre>